rename scripts to patterns
