+++ proscenium/patterns/__init__.py
