+++ proscenium/patterns/document_enricher.py
