+++ proscenium/patterns/entity_resolver.py
