+++ proscenium/patterns/knowledge_graph.py
