+++ proscenium/patterns/tools.py
